Implement big number addition and subtraction
When adding two number, it's important to evaluate the memory which
result number need. The result number may overflow if memory is
insufficient. I determine the size of allocated memory by first pick
the one which have larger allocated memory (memory size, not value).
If a's is the larger one and its most signifcant element
(unsigned long long) is not zero, I will allocate
(a->size + 1)*sizeof(unsigned long long) memory to result number.
Maybe it can use clz to check final bit, so it can use less memory if
it doesn't need it.

When subtracting two number, I use bitwise NOT and addition to
implement it.

There is a funciton bn_add_carry which is actually doing the arithmetic
. It add the element in res->num one by one, and record the carry if it
overflows.

In addition, I use bn_new, bn_znew, and bn_free to wrap allocation and
releasing function, so it can replace kmalloc with vmalloc or kvmalloc
without changing every allocation in other functions if the data size
becomes too large.